// src/legacy/upload/interface.ts
export type TUploadStatus = 'notYet' | 'uploading' | 'success' | 'error' | 'done';

export const STATUS_NOT_YET = 'notYet';
export const STATUS_UPLOADING = 'uploading';
export const STATUS_SUCCESS = 'success';
export const STATUS_ERROR = 'error';
export const STATUS_DONE = 'done';

export type TInputUploadType = 'url' | 'file';

export interface IHttpRequestHeader {
  [key: string]: string;
}

export interface IRcFile extends File {
  uid: string;
  readonly lastModifiedDate: Date;
  readonly webkitRelativePath: string;
}
// ...
export interface IUploadFile<T = any> {
  uid: string;
  size: number;
  name: string;
  fileName?: string;
  lastModified?: number;
  lastModifiedDate?: Date;
  status?: TUploadStatus;
  percent?: number;
  url?: string;
  dataUrl?: string; // 预览用的 base64
  originFile?: File | Blob; // RcUpload 返回或者说 input 返回的文件对象
  response?: T;
  error?: any;
  type: string;
  /**
   自定义文件上传失败时显示的错误信息
   */
  errorMessage?: React.ReactNode;
}